[CI] Move SampleOutput into metrics/wer.py and drop TYPE_CHECKING
A plain runtime import of SampleOutput from tasks/asr.py would be circular
because asr.py imports the WER calculators from metrics/wer.py. Moving the
dataclass down into metrics/wer.py removes the cycle, so the TYPE_CHECKING
guard and the quoted annotations go away and the metrics layer no longer
reaches up into tasks.
